pyproject.toml
[project]
name = "visite360"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.35",
    "celery>=5.4",
    "django>=5.1,<6.1",
    "django-cors-headers>=4.4",
    "django-filter>=24.3",
    "django-storages>=1.14",
    "djangorestframework>=3.16",
    "djangorestframework-simplejwt>=5.3",
    "drf-spectacular>=0.27",
    "pillow>=10.4",
    "psycopg[binary]>=3.2",
    "python-decouple>=3.8",
    "redis>=5.0",
]
